requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
pybit==5.11.0
pycryptodome==3.23.0
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
Telethon==1.40.0
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
